+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
 
-## Unreleased
+# Unreleased
+
+### Added
+
+- Add default flash notice for sign out (#178)
 
 ### Fixed